mount a fresh writable /var/run
may be a symlink to /run; that's ok because it will mount on the link
target. this is needed to prevent util-linux mount from complaining
that it can't update its gratuitous "utab" copy of the mount list.
might help with other things too, but really nothing non-root should
be trying to touch this directory anyway.